+++ application/cloudevent-converter/generic/src/main/java/org/occurrent/application/converter/generic/GenericCloudEventConverter.java
@@ -18,7 +18,10 @@ package org.occurrent.application.converter.generic;
 
 import io.cloudevents.CloudEvent;
 import org.occurrent.application.converter.CloudEventConverter;
+import org.occurrent.application.converter.typemapper.CloudEventTypeGetter;
 
+import java.util.Objects;
+import java.util.StringJoiner;
 import java.util.function.Function;
 
 /**
@@ -31,10 +34,26 @@ public class GenericCloudEventConverter<T> implements CloudEventConverter<T> {
 
     private final Function<CloudEvent, T> convertToDomainEvent;
     private final Function<T, CloudEvent> convertToCloudEvent;
+    private final CloudEventTypeGetter<T> getCloudEventType;
 
+    /**
+     * Create an instance of {@link GenericCloudEventConverter} that uses reflection (the simple name of the domain event class) as an implementation for {@link #getCloudEventType(Class)}.
+     * Use {@link GenericCloudEventConverter#GenericCloudEventConverter(Function, Function, CloudEventTypeGetter)} to configure how the cloud event type should be generated from the domain event type.
+     */
     public GenericCloudEventConverter(Function<CloudEvent, T> convertToDomainEvent, Function<T, CloudEvent> convertToCloudEvent) {
+        this(convertToDomainEvent, convertToCloudEvent, Class::getSimpleName);
+    }
+
+    /**
+     * Create an instance of {@link GenericCloudEventConverter} that uses the {@code getCloudEventType} to get the cloud event type from the domain event type.
+     */
+    public GenericCloudEventConverter(Function<CloudEvent, T> convertToDomainEvent, Function<T, CloudEvent> convertToCloudEvent, CloudEventTypeGetter<T> getCloudEventType) {
+        Objects.requireNonNull(convertToDomainEvent, "convertToDomainEvent cannot be null");
+        Objects.requireNonNull(convertToCloudEvent, "convertToCloudEvent cannot be null");
+        Objects.requireNonNull(getCloudEventType, "getCloudEventType cannot be null");
         this.convertToDomainEvent = convertToDomainEvent;
         this.convertToCloudEvent = convertToCloudEvent;
+        this.getCloudEventType = getCloudEventType;
     }
 
     @Override
@@ -45,5 +64,32 @@ public class GenericCloudEventConverter<T> implements CloudEventConverter<T> {
     @Override
     public T toDomainEvent(CloudEvent cloudEvent) {
         return convertToDomainEvent.apply(cloudEvent);
+    }
+
+    @Override
+    public String getCloudEventType(Class<? extends T> type) {
+        return getCloudEventType.getCloudEventType(type);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof GenericCloudEventConverter)) return false;
+        GenericCloudEventConverter<?> that = (GenericCloudEventConverter<?>) o;
+        return Objects.equals(convertToDomainEvent, that.convertToDomainEvent) && Objects.equals(convertToCloudEvent, that.convertToCloudEvent) && Objects.equals(getCloudEventType, that.getCloudEventType);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(convertToDomainEvent, convertToCloudEvent, getCloudEventType);
+    }
+
+    @Override
+    public String toString() {
+        return new StringJoiner(", ", GenericCloudEventConverter.class.getSimpleName() + "[", "]")
+                .add("convertToDomainEvent=" + convertToDomainEvent)
+                .add("convertToCloudEvent=" + convertToCloudEvent)
+                .add("getCloudEventType=" + getCloudEventType)
+                .toString();
     }
 }