+++ dsl/module-dsl/blocking/src/main/kotlin/se/occurrent/dsl/module/blocking/Module.kt
@@ -17,8 +17,6 @@
 package se.occurrent.dsl.module.blocking
 
 import org.occurrent.application.converter.CloudEventConverter
-import org.occurrent.application.typemapper.CloudEventTypeMapper
-import org.occurrent.application.typemapper.ReflectionCloudEventTypeMapper
 import org.occurrent.dsl.subscription.blocking.Subscriptions
 import org.occurrent.subscription.api.blocking.SubscriptionModel
 
@@ -30,10 +28,10 @@ import org.occurrent.subscription.api.blocking.SubscriptionModel
 internal annotation class ModuleDSL
 
 fun <C : Any, E : Any> module(
-    cloudEventConverter: CloudEventConverter<E>, cloudEventTypeMapper: CloudEventTypeMapper<E> = ReflectionCloudEventTypeMapper.qualified(),
+    cloudEventConverter: CloudEventConverter<E>,
     b: (@ModuleDSL ModuleBuilder<C, E>).() -> Unit
 ): Module<C> {
-    val module = ModuleBuilder<C, E>(cloudEventConverter, cloudEventTypeMapper).apply(b)
+    val module = ModuleBuilder<C, E>(cloudEventConverter).apply(b)
 
     return object : Module<C> {
         override fun dispatch(vararg commands: C) {
@@ -45,7 +43,7 @@ fun <C : Any, E : Any> module(
 }
 
 @ModuleDSL
-class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEventConverter: CloudEventConverter<E>, private val cloudEventTypeMapper: CloudEventTypeMapper<E>) {
+class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEventConverter: CloudEventConverter<E>) {
     internal val commandDispatchers = mutableListOf<CommandDispatcher<C, out Any>>()
 
     fun <B : Any> commands(commandDispatcher: CommandDispatcher<C, B>, commands: (@ModuleDSL B).() -> Unit) {
@@ -58,7 +56,7 @@ class ModuleBuilder<C : Any, E : Any> internal constructor(private val cloudEven
     }
 
     fun subscriptions(subscriptionModel: SubscriptionModel, subscriptions: (@ModuleDSL Subscriptions<E>).() -> Unit) {
-        Subscriptions(subscriptionModel, cloudEventConverter, cloudEventTypeMapper).apply(subscriptions)
+        Subscriptions(subscriptionModel, cloudEventConverter).apply(subscriptions)
     }
 }